refactor: split codewhisperer service into multiple files (#1974)
