+++ server/aws-lsp-codewhisperer/src/language-server/inline-completion/auto-trigger/editPredictionAutoTrigger.test.ts
@@ -7,7 +7,7 @@ import * as assert from 'assert'
 import * as sinon from 'sinon'
 import { editPredictionAutoTrigger } from './editPredictionAutoTrigger'
 import { EditPredictionConfigManager } from './editPredictionConfig'
-import { FileContext } from '../../../shared/codeWhispererService'
+import { FileContext } from '../../../shared/codeWhispererService/codeWhispererServiceBase'
 import { Position } from '@aws/language-server-runtimes/server-interface'
 import { CursorTracker } from '../tracker/cursorTracker'
 import { RecentEditTracker } from '../tracker/codeEditTracker'
