+++ server/aws-lsp-codewhisperer/src/language-server/securityScan/securityScanHandler.test.ts
@@ -5,7 +5,7 @@ import got from 'got'
 import * as Sinon from 'sinon'
 import { StubbedInstance, default as simon, stubInterface } from 'ts-sinon'
 import { StartCodeAnalysisRequest } from '../../client/token/codewhispererbearertokenclient'
-import { CodeWhispererServiceToken } from '../../shared/codeWhispererService'
+import { CodeWhispererServiceToken } from '../../shared/codeWhispererService/codeWhispererServiceToken'
 import { SecurityScanHandler } from './securityScanHandler'
 import { RawCodeScanIssue } from './types'
 import * as ScanConstants from './constants'
