+++ server/aws-lsp-codewhisperer/src/shared/codeWhispererService/codeWhispererServiceIAM.ts
@@ -1,0 +1,77 @@
+import {
+    CredentialsProvider,
+    CredentialsType,
+    Workspace,
+    Logging,
+    SDKInitializator,
+} from '@aws/language-server-runtimes/server-interface'
+import { AWSError, CredentialProviderChain, Credentials } from 'aws-sdk'
+import { Request } from 'aws-sdk/lib/core'
+import {
+    CodeWhispererSigv4ClientConfigurationOptions,
+    createCodeWhispererSigv4Client,
+} from '../../client/sigv4/codewhisperer'
+import CodeWhispererSigv4Client = require('../../client/sigv4/codewhisperersigv4client')
+import {
+    CodeWhispererServiceBase,
+    GenerateSuggestionsRequest,
+    GenerateSuggestionsResponse,
+    Suggestion,
+    SuggestionType,
+} from './codeWhispererServiceBase'
+
+export class CodeWhispererServiceIAM extends CodeWhispererServiceBase {
+    client: CodeWhispererSigv4Client
+    constructor(
+        credentialsProvider: CredentialsProvider,
+        workspace: Workspace,
+        logging: Logging,
+        codeWhispererRegion: string,
+        codeWhispererEndpoint: string,
+        sdkInitializator: SDKInitializator
+    ) {
+        super(codeWhispererRegion, codeWhispererEndpoint)
+        const options: CodeWhispererSigv4ClientConfigurationOptions = {
+            region: this.codeWhispererRegion,
+            endpoint: this.codeWhispererEndpoint,
+            credentialProvider: new CredentialProviderChain([
+                () => credentialsProvider.getCredentials('iam') as Credentials,
+            ]),
+        }
+        this.client = createCodeWhispererSigv4Client(options, sdkInitializator, logging)
+        // Avoid overwriting any existing client listeners
+        const clientRequestListeners = this.client.setupRequestListeners
+        this.client.setupRequestListeners = (request: Request<unknown, AWSError>) => {
+            if (clientRequestListeners) {
+                clientRequestListeners.call(this.client, request)
+            }
+            request.httpRequest.headers['x-amzn-codewhisperer-optout'] = `${!this.shareCodeWhispererContentWithAWS}`
+        }
+    }
+
+    getCredentialsType(): CredentialsType {
+        return 'iam'
+    }
+
+    async generateSuggestions(request: GenerateSuggestionsRequest): Promise<GenerateSuggestionsResponse> {
+        // add cancellation check
+        // add error check
+        if (this.customizationArn) request = { ...request, customizationArn: this.customizationArn }
+        const response = await this.client.generateRecommendations(request).promise()
+        const responseContext = {
+            requestId: response?.$response?.requestId,
+            codewhispererSessionId: response?.$response?.httpResponse?.headers['x-amzn-sessionid'],
+            nextToken: response.nextToken,
+        }
+
+        for (const recommendation of response?.recommendations ?? []) {
+            Object.assign(recommendation, { itemId: this.generateItemId() })
+        }
+
+        return {
+            suggestions: response.recommendations as Suggestion[],
+            suggestionType: SuggestionType.COMPLETION,
+            responseContext,
+        }
+    }
+}
