+++ server/aws-lsp-codewhisperer/src/shared/codeWhispererService/codeWhispererServiceToken.ts
@@ -9,167 +9,23 @@ import {
     CancellationTokenSource,
 } from '@aws/language-server-runtimes/server-interface'
 import { waitUntil } from '@aws/lsp-core/out/util/timeoutUtils'
-import { AWSError, ConfigurationOptions, CredentialProviderChain, Credentials } from 'aws-sdk'
+import { AWSError } from 'aws-sdk'
 import { PromiseResult } from 'aws-sdk/lib/request'
-import { Request } from 'aws-sdk/lib/core'
-import { v4 as uuidv4 } from 'uuid'
-import {
-    CodeWhispererSigv4ClientConfigurationOptions,
-    createCodeWhispererSigv4Client,
-} from '../client/sigv4/codewhisperer'
 import {
     CodeWhispererTokenClientConfigurationOptions,
     createCodeWhispererTokenClient,
-    RequestExtras,
-} from '../client/token/codewhisperer'
-import CodeWhispererSigv4Client = require('../client/sigv4/codewhisperersigv4client')
-import CodeWhispererTokenClient = require('../client/token/codewhispererbearertokenclient')
-import { getErrorId } from './utils'
-import { GenerateCompletionsResponse } from '../client/token/codewhispererbearertokenclient'
-
-export interface Suggestion extends CodeWhispererTokenClient.Completion, CodeWhispererSigv4Client.Recommendation {
-    itemId: string
-}
-
-export interface GenerateSuggestionsRequest extends CodeWhispererTokenClient.GenerateCompletionsRequest {
-    // TODO : This is broken due to Interface 'GenerateSuggestionsRequest' cannot simultaneously extend types 'GenerateCompletionsRequest' and 'GenerateRecommendationsRequest'.
-    //CodeWhispererSigv4Client.GenerateRecommendationsRequest {
-    maxResults: number
-}
-
-export type FileContext = GenerateSuggestionsRequest['fileContext']
-
-export interface ResponseContext {
-    requestId: string
-    codewhispererSessionId: string
-    nextToken?: string
-}
-
-export enum SuggestionType {
-    EDIT = 'EDIT',
-    COMPLETION = 'COMPLETION',
-}
-
-export interface GenerateSuggestionsResponse {
-    suggestions: Suggestion[]
-    suggestionType?: SuggestionType
-    responseContext: ResponseContext
-}
-
-// This abstract class can grow in the future to account for any additional changes across the clients
-export abstract class CodeWhispererServiceBase {
-    protected readonly codeWhispererRegion
-    protected readonly codeWhispererEndpoint
-    public shareCodeWhispererContentWithAWS = false
-    public customizationArn?: string
-    public profileArn?: string
-    abstract client: CodeWhispererSigv4Client | CodeWhispererTokenClient
-
-    inflightRequests: Set<AWS.Request<any, AWSError> & RequestExtras> = new Set()
-
-    abortInflightRequests() {
-        this.inflightRequests.forEach(request => {
-            request.abort()
-        })
-        this.inflightRequests.clear()
-    }
-
-    trackRequest(request: AWS.Request<any, AWSError> & RequestExtras) {
-        this.inflightRequests.add(request)
-    }
-
-    completeRequest(request: AWS.Request<any, AWSError> & RequestExtras) {
-        this.inflightRequests.delete(request)
-    }
-
-    abstract getCredentialsType(): CredentialsType
-
-    abstract generateSuggestions(request: GenerateSuggestionsRequest): Promise<GenerateSuggestionsResponse>
-
-    constructor(codeWhispererRegion: string, codeWhispererEndpoint: string) {
-        this.codeWhispererRegion = codeWhispererRegion
-        this.codeWhispererEndpoint = codeWhispererEndpoint
-    }
-
-    /**
-     * Updates Service Client options after client was instantiated.
-     */
-    public updateClientConfig(options: ConfigurationOptions) {
-        this.client.config.update(options)
-    }
-
-    generateItemId = () => uuidv4()
-
-    async getSubscriptionStatus(
-        statusOnly?: boolean
-    ): Promise<{ status: 'active' | 'active-expiring' | 'none'; encodedVerificationUrl?: string }> {
-        // No-op/default implementation: assume no subscription
-        return {
-            status: 'none',
-        }
-    }
-
-    async waitUntilSubscriptionActive(_cancelToken?: CancellationToken): Promise<boolean> {
-        // No-op: base class doesn't support subscription polling
-        return false
-    }
-}
-
-export class CodeWhispererServiceIAM extends CodeWhispererServiceBase {
-    client: CodeWhispererSigv4Client
-    constructor(
-        credentialsProvider: CredentialsProvider,
-        workspace: Workspace,
-        logging: Logging,
-        codeWhispererRegion: string,
-        codeWhispererEndpoint: string,
-        sdkInitializator: SDKInitializator
-    ) {
-        super(codeWhispererRegion, codeWhispererEndpoint)
-        const options: CodeWhispererSigv4ClientConfigurationOptions = {
-            region: this.codeWhispererRegion,
-            endpoint: this.codeWhispererEndpoint,
-            credentialProvider: new CredentialProviderChain([
-                () => credentialsProvider.getCredentials('iam') as Credentials,
-            ]),
-        }
-        this.client = createCodeWhispererSigv4Client(options, sdkInitializator, logging)
-        // Avoid overwriting any existing client listeners
-        const clientRequestListeners = this.client.setupRequestListeners
-        this.client.setupRequestListeners = (request: Request<unknown, AWSError>) => {
-            if (clientRequestListeners) {
-                clientRequestListeners.call(this.client, request)
-            }
-            request.httpRequest.headers['x-amzn-codewhisperer-optout'] = `${!this.shareCodeWhispererContentWithAWS}`
-        }
-    }
-
-    getCredentialsType(): CredentialsType {
-        return 'iam'
-    }
-
-    async generateSuggestions(request: GenerateSuggestionsRequest): Promise<GenerateSuggestionsResponse> {
-        // add cancellation check
-        // add error check
-        if (this.customizationArn) request = { ...request, customizationArn: this.customizationArn }
-        const response = await this.client.generateRecommendations(request).promise()
-        const responseContext = {
-            requestId: response?.$response?.requestId,
-            codewhispererSessionId: response?.$response?.httpResponse?.headers['x-amzn-sessionid'],
-            nextToken: response.nextToken,
-        }
-
-        for (const recommendation of response?.recommendations ?? []) {
-            Object.assign(recommendation, { itemId: this.generateItemId() })
-        }
-
-        return {
-            suggestions: response.recommendations as Suggestion[],
-            suggestionType: SuggestionType.COMPLETION,
-            responseContext,
-        }
-    }
-}
+} from '../../client/token/codewhisperer'
+import CodeWhispererTokenClient = require('../../client/token/codewhispererbearertokenclient')
+import { getErrorId } from '../utils'
+import { GenerateCompletionsResponse } from '../../client/token/codewhispererbearertokenclient'
+import {
+    CodeWhispererServiceBase,
+    GenerateSuggestionsRequest,
+    GenerateSuggestionsResponse,
+    ResponseContext,
+    Suggestion,
+    SuggestionType,
+} from './codeWhispererServiceBase'
 
 /**
  * Hint: to get an instance of this: `AmazonQTokenServiceManager.getInstance().getCodewhispererService()`
